+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 <a name="latest"></a>
 ### [Latest]
 
+> Features
+- **learning:** added learning section for personal use
+
+
+<a name="v4.3.2"></a>
+### [v4.3.2] - 2022-05-17
+
 > Code Refactoring
 - **packer:** removed sort on snapshots list
 
@@ -899,7 +906,8 @@
 - initial commit
 
 
-[Latest]: https://github.com/tigorlazuardi/nvim/compare/v4.3.1...HEAD
+[Latest]: https://github.com/tigorlazuardi/nvim/compare/v4.3.2...HEAD
+[v4.3.2]: https://github.com/tigorlazuardi/nvim/compare/v4.3.1...v4.3.2
 [v4.3.1]: https://github.com/tigorlazuardi/nvim/compare/v4.3.0...v4.3.1
 [v4.3.0]: https://github.com/tigorlazuardi/nvim/compare/v4.2.1...v4.3.0
 [v4.2.1]: https://github.com/tigorlazuardi/nvim/compare/v4.2.0...v4.2.1